Spymaster adds KEENEARS
KEENEARS is built in to mere courtesans but not Spymasters, who have the word 'Spy' in the job title and are expected to perform this task. Keeping abreast of court intrigue is crucial to the class roleplay. This frees a virtue for something else, like Intelligent's languages to double down on this ability, Light Step, Darkvision, etc.

Paradoxically, knowing a Spymaster (or any class) always has Keen Ears makes them overhearing conversations harder if the Spymaster does not also hide where they are. This furthers paranoia, and SS13 is nothing if not player versus paranoia.